Update the Writing Control Panel and Mission Post page
Updated the WCP to allow level 2 admins to release stuck locks. Updated the WCP to be a little simpler and utilize sub-tabs. Updated mission posts to kick over to a view page to avoid weird issues with post locks. Updated mission posting to clear the lock when a post is saved or posted.