Update the instructions for skin changes
On Site Settings, there is a block of text now that will tell admins if
they're trying to change the skins they see when they're logged in, they
need to use Site Options. On Site Options, if a user has access to the
Site Settings page, there's a block of text now that will tell admins if
they're tryin gto change the skins for visitors, they need to use Site
Settings.