Update the Writing Control Panel with more information
Hovering over the lock icon (when a post is locked) will tell you who
owns the lock and how long ago it was created. Additionally, if the lock
is more than 10 minutes old, it will automatically clear it. (It should
be cleared by that point, but in case it isn't, it'll clean the locks
up.)