Fix bug where some users could edit any post
Whoops! Turns out that the string checking we were doing wasn't very
accurate and it was causing a lot of problems with users being able
to edit posts they weren't a part of. Thanks to evshell18 on the
forums for pointing this out and getting the ball rolling on fixing it.