Rewrite the Location class
The static interface stays in place, but only as an easy way to get at
the underlying non-static methods. The order of seamless substitution
has now changed too. First, Nova will look in the override module for
a file, then in the skin, finally in the core. (Previous versions went
skin, override, core.)