Events: Add DB 'finish' event
We already have a preprocess event for DB operations that extensions
can hook into, the `prepare` event, to intervene in the operation.
However, there's also a need for an event that fires after a change
has been made to the database, including the relevant details that
follow, like insert_id or changes done.